Add more seed data and fix data_size_bytes type
- Add 7 new datasets: product_catalog, feature_store, customer_segments,
  page_views, orders, api_logs, user_activity
- Add additional dataset and column-level lineage relationships
- Fix data_size_bytes column type from Integer to BigInteger
  (to support large dataset sizes like 5GB+)
- Create migration to fix existing Integer columns to BigInteger
- Simplify initial migration by removing conditional logic
- Add example metadata values for all new datasets

+++ backend/alembic/versions/8a9b0c1d2e3f_add_dataset_metadata_fields.py
@@ -27,7 +27,7 @@ def upgrade() -> None:
     )
     op.add_column(
         'datasets',
-        sa.Column('data_size_bytes', sa.Integer(), nullable=True)
+        sa.Column('data_size_bytes', sa.BigInteger(), nullable=True)
     )
     op.add_column(
         'datasets',
